2024/07/17 - Updates

+++ SmartApparel/src/main/java/com/example/SmartApparel/Operations/repo/QuotationRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuotationRepository extends JpaRepository<Quotation, Long> {
-    // Add custom query methods if needed
+public interface QuotationRepository extends JpaRepository<Quotation, Integer> {
 }
